refactor repo, add Makefile, readme, and gitignore.
