+++ meconod/pkg/model/network/message.go
